Start extracting StackTable to header
